# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Tencent_WeKnora_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  internal_handler.DeleteTagRequest:
    properties:
      exclude_ids:
        description: Chunk seq_ids to exclude from deletion
        items:
          type: integer
        type: array
    type: object
...
info:
  contact:
    name: WeKnora Github
    url: https://github.com/Tencent/WeKnora
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
...
  /knowledge-bases/{id}/tags:
    get:
      consumes:
      - application/json
      description: 获取知识库下的所有标签及统计信息
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 标签列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 获取标签列表
      tags:
      - 标签管理
    post:
      consumes:
      - application/json
      description: 在知识库下创建新标签
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      - description: 标签信息
        in: body
        name: request
        required: true
        schema:
          properties:
            color:
              type: string
            name:
              type: string
            sort_order:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建的标签
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 创建标签
      tags:
      - 标签管理
  /knowledge-bases/{id}/tags/{tag_id}:
    delete:
      consumes:
      - application/json
      description: 删除标签，可使用force=true强制删除被引用的标签，content_only=true仅删除标签下的内容而保留标签本身
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      - description: 标签ID (UUID或seq_id)
        in: path
        name: tag_id
        required: true
        type: string
      - description: 强制删除
        in: query
        name: force
        type: boolean
      - description: 仅删除内容，保留标签
        in: query
        name: content_only
        type: boolean
      - description: 删除选项
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.DeleteTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 删除标签
      tags:
      - 标签管理
    put:
      consumes:
      - application/json
      description: 更新标签信息
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      - description: 标签ID (UUID或seq_id)
        in: path
        name: tag_id
        required: true
        type: string
      - description: 标签更新信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的标签
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 更新标签
      tags:
      - 标签管理
...
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"